handle existing symlink properly in install.py
Former-commit-id: b09b909300b4a0b7e21fc2f2b2cc8404c072ce95 [formerly 5a43555c25f79e87f23bf88faa52bc8257e273b6 [formerly e67c9bbc1ec1ea71580ffa0e9d08251bcbc6e10a] [formerly 8be8c98f6d9a4e050ee905af7589c35acd5d2345 [formerly c497908cd98fc15a5cb8c94481e2d434d835ae2d]]]
Former-commit-id: 8be8c98f6d9a4e050ee905af7589c35acd5d2345
Former-commit-id: fd3aed1ad28efc5889cf87f89cbf767ea70ee781 [formerly 35562ac163d98dedd2c14e7852a66dd217c8264c]
Former-commit-id: 9b4bb9fc68c47477c14183bf0e503884b56513f3

+++ gdsfactory/install.py
@@ -5,7 +5,6 @@ import pathlib
 import shutil
 import subprocess
 import sys
-from pathlib import Path
 
 
 def make_link(src, dest):
@@ -91,10 +90,17 @@ def copy(src: pathlib.Path, dest: pathlib.Path) -> None:
     print(f"{src} copied to {dest}")
 
 
-def make_symlink(src: Path, dest: Path) -> None:
+def make_symlink(src: pathlib.Path, dest: pathlib.Path) -> None:
     """Creates symbolic link from src to dest."""
     if dest.exists():
-        os.remove(dest)
+        if dest.is_symlink() and (src == dest.readlink()):
+            print("Nothing to do, symlink already exists:\n" + f"{src} -> {dest}")
+            return
+        print(f"removing {dest} already installed")
+        if dest.is_dir():
+            shutil.rmtree(dest)
+        else:
+            os.remove(dest)
     make_link(src, dest)
     print("Symlink made:")
     print(f"From: {src}")
